requirements.txt
-i https://pypi.org/simple
tortoise-orm[asyncpg]
fastapi
requests
passlib
python-jose[cryptography]
fastapi-jwt-auth
uvicorn[standard]
strawberry-graphql[fastapi]
tenacity
aerich
celery[redis]